+++ examples/production_beaker_react/projects/production_beaker_react-contracts/.algokit/.copier-answers.yml
@@ -1,5 +1,5 @@
 # Changes here will be overwritten by Copier; NEVER EDIT MANUALLY
-_commit: 2.1.2
+_commit: 2.1.4
 _src_path: gh:algorandfoundation/algokit-beaker-default-template
 author_email: None
 author_name: None
