+++ examples/production_python_react/projects/production_python_react-contracts/.algokit/.copier-answers.yml
@@ -1,5 +1,5 @@
 # Changes here will be overwritten by Copier; NEVER EDIT MANUALLY
-_commit: 1.2.2
+_commit: 1.2.4
 _src_path: gh:algorandfoundation/algokit-python-template
 author_email: None
 author_name: None
