+++ examples/production_python_react/projects/production_python_react-contracts/.algokit/generators/create_contract/smart_contracts/{{ contract_name }}/deploy_config.py.j2
@@ -14,7 +14,7 @@ def deploy(
     app_spec: algokit_utils.ApplicationSpecification,
     deployer: algokit_utils.Account,
 ) -> None:
-    from smart_contracts.artifacts.{{ contract_name }}.client import (
+    from smart_contracts.artifacts.{{ contract_name }}.{{ contract_name }}_client import (
         {{ contract_name.split('_')|map('capitalize')|join }}Client,
     )
