+++ examples/production_python_react/projects/production_python_react-contracts/smart_contracts/__main__.py
@@ -4,10 +4,9 @@ from pathlib import Path
 
 from dotenv import load_dotenv
 
-from smart_contracts.config import contracts
-from smart_contracts.helpers.build import build
-from smart_contracts.helpers.deploy import deploy
-from smart_contracts.helpers.util import find_app_spec_file
+from smart_contracts._helpers.build import build
+from smart_contracts._helpers.config import contracts
+from smart_contracts._helpers.deploy import deploy
 
 # Uncomment the following lines to enable auto generation of AVM Debugger compliant sourcemap and simulation trace file.
 # Learn more about using AlgoKit AVM Debugger to debug your TEAL source codes and inspect various kinds of
@@ -36,7 +35,14 @@ def main(action: str) -> None:
             for contract in contracts:
                 logger.info(f"Deploying app {contract.name}")
                 output_dir = artifact_path / contract.name
-                app_spec_file_name = find_app_spec_file(output_dir)
+                app_spec_file_name = next(
+                    (
+                        file.name
+                        for file in output_dir.iterdir()
+                        if file.is_file() and file.suffixes == [".arc32", ".json"]
+                    ),
+                    None,
+                )
                 if app_spec_file_name is None:
                     raise Exception("Could not deploy app, .arc32.json file not found")
                 app_spec_path = output_dir / app_spec_file_name
