+++ examples/production_python_react/projects/production_python_react-contracts/smart_contracts/_helpers/config.py
