--- examples/production_python_react/projects/production_python_react-contracts/smart_contracts/helpers/util.py
@@ -1,8 +1,0 @@
-from pathlib import Path
-
-
-def find_app_spec_file(output_dir: Path) -> str | None:
-    for file in output_dir.iterdir():
-        if file.is_file() and file.suffixes == [".arc32", ".json"]:
-            return file.name
-    return None
