+++ examples/starter_beaker_react/projects/starter_beaker_react-contracts/smart_contracts/_helpers/config.py
