+++ examples/starter_python_react/projects/starter_python_react-contracts/smart_contracts/_helpers/config.py
